Setting up scheduler tests.
